apis/brotli-enc/requirements.txt
azure-functions

brotli
